Exceptions resourceNotFound e DatabaseException

+++ src/main/java/com/estudo/desafio/service/ClientService.java
@@ -3,10 +3,15 @@ package com.estudo.desafio.service;
 import com.estudo.desafio.dto.ClientDTO;
 import com.estudo.desafio.entities.Client;
 import com.estudo.desafio.repository.ClientRepository;
+import com.estudo.desafio.service.exception.DatabaseException;
+import com.estudo.desafio.service.exception.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.swing.text.html.parser.Entity;
@@ -22,9 +27,8 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id) {
-        Optional<Client> obj = repository.findById(id);
-        Client client = obj.get();
-        return new ClientDTO(client);
+        Client obj = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Cliente não encontrado"));
+        return new ClientDTO(obj);
 
 
     }
@@ -52,21 +56,31 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto) {
-
+        try {
         Client entity = repository.getReferenceById(id);
         copyDtoEntity(dto, entity);
         entity = repository.save(entity);
         return new ClientDTO(entity);
 
+        }catch(EntityNotFoundException e) {
+            throw new ResourceNotFoundException("cliente não encontrado");
+        }
     }
 
-
-
-
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("client não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
+
     private void copyDtoEntity(ClientDTO dto, Client entity) {
         entity.setName(dto.getName());
         entity.setCpf(dto.getCpf());
